añadiendo opcion de bloqueo/parry
